# docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  auth.CreateAccessTokenRequest:
    properties:
      password:
        example: <password>
        type: string
      user_name:
        example: <user name>
        type: string
    required:
    - password
    - user_name
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      type:
        type: string
    type: object
  common.APIError:
    properties:
      message:
        type: string
    type: object
  common.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/common.APIError'
        type: array
    type: object
...
host: '{{.Host}}'
info:
  contact:
    email: gyuhwan.a.kim@gmail.com
    name: API Support
  description: This is a sample gin starter server.
  license:
    name: MIT
    url: https://github.com/gghcode/go-gin-starterkit/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Go Gin Starter API
  version: "1.0"
paths:
...
  /api/oauth2/token:
    post:
      consumes:
      - application/json
      description: Get new access token
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.CreateAccessTokenRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/auth.TokenResponse'
            type: object
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/common.ErrorResponse'
            type: object
        "401":
          description: Invalid credential
          schema:
            $ref: '#/definitions/common.ErrorResponse'
            type: object
      tags:
      - Auth API
...
securityDefinitions:
  "":
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"